+++ src/photon/tests/pathTest.php
@@ -25,26 +25,17 @@
 
 namespace photon\tests\pathTest;
 
+use \photon\test\TestCase;
 use \photon\config\Container as Conf;
 use \photon\path\Dir;
 
-class PathTest extends \PHPUnit_Framework_TestCase
+class PathTest extends TestCase
 {
-    protected $conf;
-
-    public function setUp()
-    {
-        $this->conf = Conf::dump();
-    }
-
-    public function tearDown()
-    {
-        Conf::load($this->conf);
-    }
-
     public function testListFiles()
     {
         // review this test, adding a file in the data folder breaks it...
-        $this->assertEquals(13, count(Dir::listFiles(__DIR__ . '/../data')));
+        // Same if we remove a file... We really need to do somethings !
+        // It's can wait the 2.x series
+        $this->assertEquals(1, count(Dir::listFiles(__DIR__ . '/../data')));
     }
 }